tests: Move bot storage and config tests to test_embedded_bot_system.py.
TestServiceBotStateHandler and TestServiceBotConfigHandler cover the
bot storage and bot config infrastructure, which only embedded bots
use, rather than the message send codepaths that the rest of
test_service_bot_system.py covers.

This is a prep commit that lets the next commit rename
test_service_bot_system.py after what it is left testing.
